+++ packages/ionic-react/CHANGELOG.md
@@ -8,6 +8,7 @@
 - remove `style` option from the `application` schematic (now defaults to CSS)
 - remove `pascalCaseComponent` option from the `application` schematic (now defaults to true)
 - remove `skipFormat` option from the `application` schematic (now defaults to false)
+- remove `linter` option from the `application` schematic (now defaults to ESLint)
 
 # 4.1.0